server/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
